+++ web/src/app/browser/src/defaultBrowserRules.ts
@@ -2,9 +2,9 @@ import { ModifierKeyConstants } from '@keymanapp/common-types';
 import {
   Codes,
   DefaultRules,
-  type KeyEvent,
-  type OutputTarget
+  type KeyEvent
 } from 'keyman/engine/keyboard';
+import { type OutputTarget } from 'keyman/engine/js-processor';
 
 import ContextManager from './contextManager.js';
