+++ web/src/engine/element-wrappers/src/outputTarget.ts
@@ -1,4 +1,4 @@
-import { OutputTarget as OutputTargetBase } from "keyman/engine/keyboard";
+import { OutputTarget as OutputTargetBase } from "keyman/engine/js-processor";
 import { EventEmitter } from 'eventemitter3';
 
 export default abstract class OutputTarget<EventMap extends EventEmitter.ValidEventTypes> extends OutputTargetBase {
