+++ web/src/engine/interfaces/src/prediction/predictionContext.ts
@@ -1,7 +1,6 @@
 import { EventEmitter } from "eventemitter3";
 import { type LanguageProcessorSpec , ReadySuggestions, type InvalidateSourceEnum, StateChangeHandler } from './languageProcessor.interface.js';
-import { type OutputTarget } from "keyman/engine/keyboard";
-import { type KeyboardProcessor } from 'keyman/engine/js-processor';
+import { type KeyboardProcessor, type OutputTarget } from 'keyman/engine/js-processor';
 
 interface PredictionContextEventMap {
   update: (suggestions: Suggestion[]) => void;
