+++ web/src/engine/js-processor/src/index.ts
@@ -4,3 +4,8 @@ export { default as RuleBehavior } from "./ruleBehavior.js";
 export * from './kbdInterface.js';
 export { default as KeyboardInterface } from "./kbdInterface.js";
 export * from "./systemStores.js";
+export * from "./deadkeys.js";
+export { default as OutputTarget } from "./outputTarget.js";
+export * from "./outputTarget.js";
+export { Mock } from "./mock.js";
+export * from "./stringDivergence.js";
