+++ web/src/engine/main/src/headless/contextWindow.ts
@@ -1,4 +1,4 @@
-import { Mock } from "keyman/engine/keyboard";
+import { Mock } from "keyman/engine/js-processor";
 
 export default class ContextWindow implements Context {
   // Used to limit the range of context replicated for use of keyboard rules within
