+++ web/src/test/auto/headless/engine/interfaces/prediction/predictionContext.spec.js
@@ -4,8 +4,8 @@ import sinon from 'sinon';
 import { LanguageProcessor, TranscriptionCache } from 'keyman/engine/main';
 import { PredictionContext } from 'keyman/engine/interfaces';
 import { Worker as LMWorker } from "@keymanapp/lexical-model-layer/node";
-import { DeviceSpec, Mock } from 'keyman/engine/keyboard';
-import { KeyboardProcessor } from 'keyman/engine/js-processor';
+import { DeviceSpec } from 'keyman/engine/keyboard';
+import { KeyboardProcessor, Mock } from 'keyman/engine/js-processor';
 
 function compileDummyModel(suggestionSets) {
   return `
