+++ web/src/tools/testing/recorder-core/src/nodeProctor.ts
@@ -8,7 +8,8 @@ import {
   RecordedSyntheticKeystroke
 } from "./index.js";
 
-import { KeyEvent, KeyEventSpec, Mock, type OutputTarget, KeyboardHarness } from "keyman/engine/keyboard";
+import { KeyEvent, KeyEventSpec, KeyboardHarness } from "keyman/engine/keyboard";
+import { Mock, type OutputTarget } from "keyman/engine/js-processor";
 import { DeviceSpec } from "@keymanapp/web-utils";
 import { KeyboardInterface, KeyboardProcessor } from 'keyman/engine/js-processor';
