+++ web/src/tools/testing/recorder/browserProctor.ts
@@ -4,7 +4,7 @@
 
 import { type DeviceSpec } from "@keymanapp/web-utils";
 
-import { type OutputTarget } from "keyman/engine/keyboard";
+import { type OutputTarget } from "keyman/engine/js-processor";
 
 import { type KeymanEngine } from 'keyman/app/browser';
